Cleanup connections (#1725)
* cleanup connecting to CRI client


* Fix setCookies tests


* Add client close on cleanup


* Make error message assertion case insensitive
