+++ lib/handlers/domHandler.js
@@ -139,6 +139,13 @@ const calculateNewCenter = async (sourceElemobjectId, destElem) => {
   };
 };
 
+const setFileInputFiles = async (objectId, resolvedPath) => {
+  await dom.setFileInputFiles({
+    objectId,
+    files: [resolvedPath],
+  });
+};
+
 module.exports = {
   boundBox,
   boundingBoxCenter,
@@ -146,4 +153,5 @@ module.exports = {
   getPositionalDifference,
   getBoxModel,
   calculateNewCenter,
+  setFileInputFiles,
 };
