+++ lib/handlers/networkHandler.js
@@ -78,7 +78,28 @@ const setNetworkEmulation = async (networkType) => {
   });
 };
 
+const setUserAgent = async (deviceEmulate) => {
+  await network.setUserAgentOverride({
+    userAgent: deviceEmulate.userAgent,
+  });
+};
+
+const setCookie = async (options) => await network.setCookie(options);
+
+const getCookies = async (options) => {
+  return await network.getCookies(options);
+};
+
+const clearBrowserCookies = async () => await network.clearBrowserCookies();
+
+const deleteCookies = async (options) => await network.deleteCookies(options);
+
 module.exports = {
   setNetworkEmulation,
   resetPromises,
+  setUserAgent,
+  setCookie,
+  getCookies,
+  clearBrowserCookies,
+  deleteCookies,
 };
