+++ lib/handlers/pageHandler.js
@@ -79,6 +79,13 @@ const addJavascriptDialogOpeningListener = () => {
   });
 };
 
+const handleJavaScriptDialog = async (accept, prompText) => {
+  await page.handleJavaScriptDialog({
+    accept,
+    prompText,
+  });
+};
+
 const resetPromises = () => {
   frameNavigationPromise = {};
   framePromises = {};
@@ -230,9 +237,24 @@ const captureScreenshot = async function (domHandler, selector, options) {
   return screenShot;
 };
 
+const closePage = async () => page.close();
+
+const reload = async (value) => page.reload({ ignoreCache: value });
+
+const getNavigationHistory = async () => page.getNavigationHistory();
+
+const navigateToHistoryEntry = async (entryId) => {
+  await page.navigateToHistoryEntry({ entryId });
+};
+
 module.exports = {
   handleNavigation,
   resetPromises,
   captureScreenshot,
   addJavascriptDialogOpeningListener,
+  closePage,
+  reload,
+  navigateToHistoryEntry,
+  handleJavaScriptDialog,
+  getNavigationHistory,
 };
