+++ lib/taiko.js
@@ -1,6 +1,4 @@
 const { doActionAwaitingNavigation } = require('./doActionAwaitingNavigation');
-const cri = require('chrome-remote-interface');
-const isReachable = require('is-reachable');
 const {
   wait,
   isString,
@@ -12,7 +10,6 @@ const {
   isSelector,
   isElement,
   isObject,
-  isPromise,
 } = require('./helper');
 const inputHandler = require('./handlers/inputHandler');
 const domHandler = require('./handlers/domHandler');
@@ -24,7 +21,6 @@ const runtimeHandler = require('./handlers/runtimeHandler');
 const browserHandler = require('./handlers/browserHandler');
 const emulationHandler = require('./handlers/emulationHandler');
 const { description } = require('./actions/pageActionChecks');
-const { logEvent } = require('./logger');
 const { match, $$xpath, findElements, findFirstElement } = require('./elementSearch');
 const { handleRelativeSearch, RelativeSearchElement } = require('./proximityElementSearch');
 const Element = require('./elements/element');
@@ -44,160 +40,18 @@ const crypto = require('crypto');
 const { eventHandler, eventRegexMap } = require('./eventBus');
 const overlayHandler = require('./handlers/overlayHandler');
 const { highlightElement } = require('./elements/elementHelper');
-const numRetries = defaultConfig.criConnectionRetries;
+const { launchBrowser } = require('./browserLauncher');
 const {
-  launchBrowser,
-  closeBrowser,
-  errorMessageForBrowserProcessCrash,
-} = require('./browserLauncher');
-const { getPlugins, registerHooks, pluginHooks } = require('./plugins');
-let page,
-  network,
-  input,
-  _client,
-  dom,
-  overlay,
-  security,
-  device,
-  eventHandlerProxy,
-  clientProxy,
-  connectedToRemoteBrowser = false;
+  connect_to_cri,
+  closeConnection,
+  cleanUpListenersOnClient,
+  validate,
+  getClient,
+} = require('./connection');
+const { getPlugins, registerHooks } = require('./plugins');
+let eventHandlerProxy;
 
 module.exports.emitter = descEvent;
-
-const createProxyForCDPDomain = (cdpClient, cdpDomainName) => {
-  const cdpDomain = cdpClient[cdpDomainName];
-  const cdpDomainProxy = new Proxy(cdpDomain, {
-    get: (target, name) => {
-      const domainApi = target[name];
-      if (typeof domainApi === 'function') {
-        return async (...args) => {
-          return await new Promise((resolve, reject) => {
-            eventHandler.removeAllListeners('browserCrashed');
-            eventHandler.on('browserCrashed', () => {
-              _client.removeAllListeners();
-              clientProxy = null;
-              reject();
-            });
-            let res = domainApi.apply(null, args);
-            if (isPromise(res)) {
-              res.then(resolve).catch(reject);
-            } else {
-              resolve(res);
-            }
-          }).catch((e) => {
-            if (e.message.match(/WebSocket is not open: readyState 3/)) {
-              errorMessageForBrowserProcessCrash();
-            }
-            throw e;
-          });
-        };
-      } else {
-        return domainApi;
-      }
-    },
-  });
-  cdpClient[cdpDomainName] = cdpDomainProxy;
-  return cdpClient[cdpDomainName];
-};
-const initCRIProperties = (c) => {
-  page = createProxyForCDPDomain(c, 'Page');
-  network = createProxyForCDPDomain(c, 'Network');
-  createProxyForCDPDomain(c, 'Runtime');
-  input = createProxyForCDPDomain(c, 'Input');
-  dom = createProxyForCDPDomain(c, 'DOM');
-  overlay = createProxyForCDPDomain(c, 'Overlay');
-  security = createProxyForCDPDomain(c, 'Security');
-  createProxyForCDPDomain(c, 'Browser');
-  createProxyForCDPDomain(c, 'Target');
-  createProxyForCDPDomain(c, 'Emulation');
-  _client = c;
-  clientProxy = getEventProxy(_client);
-};
-
-const initCRI = async (target, n, options = {}) => {
-  try {
-    ({ target, options } = await pluginHooks.preConnectionHook(target, options));
-    var c = await cri({
-      target,
-      host: defaultConfig.host,
-      port: defaultConfig.port,
-      alterPath: defaultConfig.alterPath,
-      local: defaultConfig.local,
-    });
-    const promises = [];
-    eventHandler.on('handlerActingOnNewSession', (promise) => {
-      promises.push(promise);
-    });
-    initCRIProperties(c);
-    const domainEnablePromises = [network.enable(), page.enable(), dom.enable(), security.enable()];
-    if (!defaultConfig.firefox) {
-      domainEnablePromises.push(overlay.enable());
-    }
-    await Promise.all(domainEnablePromises);
-    _client.on('disconnect', reconnect);
-    // Should be emitted after enabling all domains. All handlers can then perform any action on domains properly.
-    eventHandler.emit('createdSession', _client, target);
-    if (defaultConfig.ignoreSSLErrors) {
-      security.setIgnoreCertificateErrors({ ignore: true });
-    }
-    device = process.env.TAIKO_EMULATE_DEVICE;
-    if (device) {
-      emulateDevice(device);
-    }
-    await Promise.all(promises);
-    eventHandler.removeAllListeners('handlerActingOnNewSession');
-    logEvent('Session Created');
-    return c;
-  } catch (error) {
-    logEvent(error);
-    if (n < 2) {
-      throw error;
-    }
-    return new Promise((r) => setTimeout(r, 100)).then(
-      async () => await initCRI(target, n - 1, options),
-    );
-  }
-};
-
-const connect_to_cri = async (target, options = {}) => {
-  if (_client && _client._ws.readyState === 1) {
-    if (!defaultConfig.firefox) {
-      await network.setRequestInterception({
-        patterns: [],
-      });
-    }
-    _client.removeAllListeners();
-  }
-  var tgt = target || (await targetHandler.waitForTargetToBeCreated(numRetries));
-  return initCRI(tgt, numRetries, options);
-};
-
-async function reconnect() {
-  const response = await isReachable(`${defaultConfig.host}:${defaultConfig.port}`);
-  if (response) {
-    try {
-      logEvent('Reconnecting');
-      eventHandler.emit('reconnecting');
-      _client.removeAllListeners();
-      const pages = await targetHandler.getFirstAvailablePageTarget();
-      await connect_to_cri(pages[0].targetId);
-      logEvent('Reconnected');
-      eventHandler.emit('reconnected');
-    } catch (e) {}
-  }
-}
-
-eventHandler.addListener('targetCreated', async (newTarget) => {
-  const response = await isReachable(`${defaultConfig.host}:${defaultConfig.port}`);
-  if (response) {
-    const pages = await targetHandler.getFirstAvailablePageTarget();
-    await connect_to_cri(pages[0].targetId).then(() => {
-      logEvent(`Target Navigated: Target id: ${newTarget.targetInfo.targetId}`);
-      eventHandler.emit('targetNavigated');
-    });
-  }
-});
 
 /**
  * Launches a browser with a tab. The browser will be closed when the parent node.js process is closed.<br>
@@ -245,7 +99,7 @@ module.exports.openBrowser = async (
     defaultConfig.host = options.host;
     defaultConfig.port = options.port;
     defaultConfig.browserDebugUrl = options.target;
-    connectedToRemoteBrowser = true;
+    defaultConfig.connectedToRemoteBrowser = true;
   } else {
     const { currentHost, currentPort, browserDebugUrl } = await launchBrowser(options);
     defaultConfig.host = currentHost;
@@ -253,7 +107,9 @@ module.exports.openBrowser = async (
     defaultConfig.browserDebugUrl = browserDebugUrl;
   }
   await connect_to_cri();
-  var description = device ? `Browser opened with viewport ${device}` : 'Browser opened';
+  var description = defaultConfig.device
+    ? `Browser opened with viewport ${defaultConfig.device}`
+    : 'Browser opened';
   descEvent.emit('success', description);
 
   if (process.env.TAIKO_EMULATE_NETWORK) {
@@ -282,33 +138,14 @@ module.exports.closeBrowser = async () => {
 };
 
 const _closeBrowser = async () => {
-  let timeout;
   fetchHandler.resetInterceptors();
-  if (_client) {
-    // remove listeners other than JS dialog for beforeUnload on client first to stop executing them when closing
-    await _client.removeAllListeners();
-    pageHandler.addJavascriptDialogOpeningListener();
-    //TODO: Remove check once fixed https://bugs.chromium.org/p/chromium/issues/detail?id=1147809
-    if (!defaultConfig.firefox && !(process.platform == 'win32' && defaultConfig.headful)) {
-      await page.close();
-
-      await new Promise((resolve) => {
-        timeout = setTimeout(() => {
-          resolve();
-        }, defaultConfig.retryTimeout);
-        Promise.all(promisesToBeResolvedBeforeCloseBrowser).then(() => {
-          clearTimeout(timeout);
-          resolve();
-        });
-      });
-    }
-    clientProxy = null;
-  }
-  connectedToRemoteBrowser ? await _client.Browser.close() : await closeBrowser();
-  await _client.close();
+  await closeConnection(promisesToBeResolvedBeforeCloseBrowser);
 };
 
 function getEventProxy(target) {
+  if (!target) {
+    return target;
+  }
   let unsupportedClientMethods = [
     'removeListener',
     'emit',
@@ -333,7 +170,7 @@ function getEventProxy(target) {
  *
  * @returns {Object}
  */
-module.exports.client = () => clientProxy;
+module.exports.client = () => getEventProxy(getClient());
 
 /**
  * Allows switching between tabs and windows using URL or page title or Window name.
@@ -491,9 +328,7 @@ async function emulateDevice(deviceModel) {
   }
   await Promise.all([
     emulationHandler.setViewport(deviceEmulate.viewport),
-    network.setUserAgentOverride({
-      userAgent: deviceEmulate.userAgent,
-    }),
+    networkHandler.setUserAgent(deviceEmulate),
   ]);
   descEvent.emit('success', 'Device emulation set to ' + deviceModel);
 }
@@ -572,7 +407,7 @@ module.exports.openTab = async (targetUrl, options = {}) => {
   }
 
   const createNewTarget = async () => {
-    _client.removeAllListeners();
+    await cleanUpListenersOnClient();
     let target = await targetHandler.createTarget(targetUrl);
     await connect_to_cri(target);
     targetHandler.register(options.name, target);
@@ -621,7 +456,7 @@ module.exports.openIncognitoWindow = async (url, options = {}) => {
     url = 'http://' + url;
   }
   const targetId = await targetHandler.createBrowserContext(url, options);
-  _client.removeAllListeners();
+  await cleanUpListenersOnClient();
   await connect_to_cri(targetId);
   targetHandler.register(options.name, targetId);
 
@@ -709,8 +544,7 @@ module.exports.closeTab = async (identifier) => {
     targetHandler.isMatchingRegex(activeTab, identifier) ||
     targetHandler.isMatchingTarget(activeTab, identifier)
   ) {
-    _client.removeAllListeners();
-    await _client.close();
+    await cleanUpListenersOnClient();
     await connect_to_cri(others[0].targetId);
   }
   let message = identifier
@@ -783,7 +617,7 @@ module.exports.setCookie = async (name, value, options = {}) => {
   }
   options.name = name;
   options.value = value;
-  let res = await network.setCookie(options);
+  let res = await networkHandler.setCookie(options);
   if (!res.success) {
     throw new Error('Unable to set ' + name + ' cookie');
   }
@@ -811,14 +645,14 @@ module.exports.setCookie = async (name, value, options = {}) => {
 module.exports.deleteCookies = async (cookieName, options = {}) => {
   validate();
   if (!cookieName || !cookieName.trim()) {
-    await network.clearBrowserCookies();
+    await networkHandler.clearBrowserCookies();
     descEvent.emit('success', 'Browser cookies deleted successfully');
   } else {
     if (options.url === undefined && options.domain === undefined) {
       throw new Error('At least URL or domain needs to be specified for deleting cookies');
     }
     options.name = cookieName;
-    await network.deleteCookies(options);
+    await networkHandler.deleteCookies(options);
     descEvent.emit('success', `"${cookieName}" cookie deleted successfully`);
   }
 };
@@ -838,7 +672,7 @@ module.exports.deleteCookies = async (cookieName, options = {}) => {
  */
 module.exports.getCookies = async (options = {}) => {
   validate();
-  return (await network.getCookies(options)).cookies;
+  return (await networkHandler.getCookies(options)).cookies;
 };
 
 /**
@@ -942,7 +776,7 @@ module.exports.reload = async (
   options = setNavigationOptions(options);
   await doActionAwaitingNavigation(options, async () => {
     const value = options.ignoreCache || false;
-    await page.reload({ ignoreCache: value });
+    await pageHandler.reload(value);
   });
   let windowLocation = (await runtimeHandler.runtimeEvaluate('window.location.toString()')).result
     .value;
@@ -992,7 +826,7 @@ module.exports.goForward = async (
 };
 
 const _go = async (delta, options) => {
-  const history = await page.getNavigationHistory();
+  const history = await pageHandler.getNavigationHistory();
   const entry = history.entries[history.currentIndex + delta];
   if (!entry) {
     return null;
@@ -1005,7 +839,7 @@ const _go = async (delta, options) => {
   }
   options = setNavigationOptions(options);
   await doActionAwaitingNavigation(options, async () => {
-    await page.navigateToHistoryEntry({ entryId: entry.id });
+    await pageHandler.navigateToHistoryEntry(entry.id);
   });
 };
 
@@ -1184,7 +1018,7 @@ module.exports.hover = async (selector, options = {}) => {
     Promise.resolve()
       .then(() => {
         option.type = 'mouseMoved';
-        return input.dispatchMouseEvent(option);
+        return inputHandler.dispatchMouseEvent(option);
       })
       .catch((err) => {
         throw new Error(err);
@@ -1414,10 +1248,7 @@ module.exports.attach = async (filepath, to) => {
   if (defaultConfig.headful) {
     await highlightElement(element);
   }
-  await dom.setFileInputFiles({
-    objectId: element.get(),
-    files: [resolvedPath],
-  });
+  await domHandler.setFileInputFiles(element.get(), resolvedPath);
   descEvent.emit('success', 'Attached ' + resolvedPath + ' to the ' + description(to, true));
 };
 
@@ -1612,7 +1443,7 @@ async function mouseAction(selector, action, coordinates, options = {}) {
     } else if (action === 'release') {
       options.type = 'mouseReleased';
     }
-    await input.dispatchMouseEvent(options);
+    await inputHandler.dispatchMouseEvent(options);
   });
   descEvent.emit(
     'success',
@@ -2809,10 +2640,7 @@ module.exports.waitFor = waitFor;
  * @return {Promise<void>} - Resolves when dialog is accepted
  */
 module.exports.accept = async (text = '') => {
-  await page.handleJavaScriptDialog({
-    accept: true,
-    promptText: text,
-  });
+  await pageHandler.handleJavaScriptDialog(true, text);
   descEvent.emit('success', 'Accepted dialog');
 };
 
@@ -2825,9 +2653,7 @@ module.exports.accept = async (text = '') => {
  * @return {Promise<void>} - Resolves when dialog is dismissed
  */
 module.exports.dismiss = async () => {
-  await page.handleJavaScriptDialog({
-    accept: false,
-  });
+  await pageHandler.handleJavaScriptDialog(false);
   descEvent.emit('success', 'Dismissed dialog');
 };
 
@@ -3004,18 +2830,6 @@ const evaluate = async (selector, func) => {
   let objectId = (await findFirstElement(selector)).get();
   let { result } = await runtimeHandler.runtimeCallFunctionOn(func, null, { objectId: objectId });
   return result || {};
-};
-
-const validate = () => {
-  if (!_client) {
-    throw new Error('Browser or page not initialized. Call `openBrowser()` before using this API');
-  }
-  if (_client._ws.readyState > 1) {
-    errorMessageForBrowserProcessCrash();
-    throw new Error(
-      "Connection to browser lost. This probably isn't a problem with Taiko, inspect logs for possible causes.",
-    );
-  }
 };
 
 const rectangle = async (selector, callback) => {
